fix: add hook and fix lint errors
