+++ src/script/media/MediaStreamErrorTypes.ts
@@ -31,15 +31,13 @@ export const MEDIA_STREAM_ERROR_TYPES = {
 };
 
 export function isMediaStreamDeviceError(errorStr: string): boolean {
-  return MEDIA_STREAM_ERROR_TYPES.DEVICE.includes(errorStr);
+  return MEDIA_STREAM_ERROR_TYPES.DEVICE.includes(errorStr as MEDIA_STREAM_ERROR);
 }
 
 export function isMediaStreamReadDeviceError(errorStr: string): boolean {
-  return (
-    [
-      MEDIA_STREAM_ERROR.NOT_READABLE_ERROR,
-      MEDIA_STREAM_ERROR.NOT_ALLOWED_ERROR,
-      MEDIA_STREAM_ERROR.NOT_FOUND_ERROR,
-    ].includes(errorStr)
-  );
+  return [
+    MEDIA_STREAM_ERROR.NOT_READABLE_ERROR,
+    MEDIA_STREAM_ERROR.NOT_ALLOWED_ERROR,
+    MEDIA_STREAM_ERROR.NOT_FOUND_ERROR,
+  ].includes(errorStr as MEDIA_STREAM_ERROR);
 }
